refactor(analysis-types): move license serde tests
