+++ crates/tokmd-analysis-types/src/license.rs
@@ -20,3 +20,18 @@ pub enum LicenseSourceKind {
     Metadata,
     Text,
 }
+
+#[cfg(test)]
+mod tests {
+    use super::LicenseSourceKind;
+
+    #[test]
+    fn license_source_kind_serde_roundtrip() -> Result<(), Box<dyn std::error::Error>> {
+        for variant in [LicenseSourceKind::Metadata, LicenseSourceKind::Text] {
+            let json = serde_json::to_string(&variant)?;
+            let back: LicenseSourceKind = serde_json::from_str(&json)?;
+            assert_eq!(back, variant);
+        }
+        Ok(())
+    }
+}
